// src/app/models/helpdesk/claim.model.ts
import { Resource } from '../shared/ressource.model';
import { Warehouse } from '../inventory/warehouse.model';
import { ClaimType } from './claim-type.model';
import { ClaimStatus } from './claim-status.model';
import { ReducedDocument } from '../sales/reduced-document.model';
import { ReducedDocumentLine } from '../sales/reduced-document-line.model';
import { ReducedTiers } from '../achat/reduced-tiers.model';
import { ReducedContact } from '../shared/reduced-contact.model';
import { ReducedItem } from '../inventory/reduced-item.model';

export class Claim extends Resource {
  Code: string;
  ClaimType: string;
  Description: string;
  Informations: string;
  Reference: string;
  DocumentTypeCode: string;
  IsTreated: boolean;
  IdWarehouse: number;
  IdWarehouseNavigation: Warehouse;
  IdItem: number;
  IdItemNavigation: ReducedItem;
  IdDocument: number;
  IdDocumentNavigation: ReducedDocument;
  IdPurchaseDocument: number;
  IdPurchaseDocumentNavigation: ReducedDocument;
  IdSalesDocument: number;
  IdSalesDocumentNavigation: ReducedDocument;
  IdReceiptDocument: number;
  IdReceiptDocumentNavigation: ReducedDocument;
  IdDeliveryDocument: number;
  IdDeliveryDocumentNavigation: ReducedDocument;
  IdDocumentLine: number;
  IdDocumentLineNavigation: ReducedDocumentLine;
  IdClient: number;
  IdClientNavigation: ReducedTiers;
  IdContact: number;
  IdContactNavigation: ReducedContact;
  IdFournisseur: number;
  IdFournisseurNavigation: ReducedTiers;
  DocumentDate: Date;
  ValidationDate: Date;
  IdClaimStatus:  number;
  IdClaimStatusNavigation: ClaimStatus;
  ClaimTypeNavigation: ClaimType;
  ClaimQty: number;
  ClaimMaxQty: number;
  ClaimInteraction: any;
  StockMovement: any;
  IdSalesAssetDocument: number;
  IdSalesAssetDocumentNavigation: ReducedDocument;
  IdPurchaseAssetDocument: number;
  IdPurchaseAssetDocumentNavigation: ReducedDocument;
  IsClaimQtyLocked: boolean;
  IdMovementIn: number;
  IdMovementOut: number;
  ReferenceOldDocument: string;
  IdMovementInNavigation: ReducedDocument;
  IdMovementOutNavigation: ReducedDocument;
}